requirements.txt
diffusers==0.30.3
gradio==5.7.1
numpy==2.1.3
opencv_python==4.9.0.80
opencv_python_headless==4.10.0.84
pandas==2.2.3
Pillow==11.0.0
PyYAML==6.0.1
PyYAML==6.0.2
scikit_learn==1.5.2
torch==2.0.1
torch_xla==2.5.1
torchmetrics==1.3.1
torchvision==0.15.2
tqdm==4.66.4
transformers==4.44.2
